assign/Assistant_api/PromptEvaluator/requirements.txt

openai==0.28
python-dotenv>=1.0
tqdm>=4.66


